src/plugins/paimon_info/__init__.py
from typing import TYPE_CHECKING

from khl import Message

from ...database.models.cookie import PrivateCookie

if TYPE_CHECKING:
    from ...bot import LittlePaimon


async def on_startup(bot: 'LittlePaimon'):
    @bot.my_command('ys', aliases=['原神卡片', '个人卡片'])
    async def ys(msg: Message, uid: str = None):
        if uid is None:
            cookies = await PrivateCookie.filter(user_id=msg.author.id)
            if len(cookies) == 0:
                await msg.reply('你还没有绑定 cookie')
            else:
                ...
        else:
            cookies = await PrivateCookie.filter(uid=uid)
            if len(cookies) == 0:
                await msg.reply(f'数据库没有查询到 uid 的信息')
                return
            for ck in cookies:
                if ck.user_id != msg.author.id:
                    await msg.reply(f'UID: {ck.uid} 不属于你')
                else:
                    ...
