requirements.txt
khl.py>=0.3.2
colorlog>=6.6.0
colorama>=0.4.5
pydantic>=1.9.2
gevent>=21.12.0
Flask>=2.2.2
pywebio>=1.6.2
httpx>=0.23.0
Pillow>=9.2.0
tqdm>=4.64.0
KaiheilaCardBuilder>=1.2.5
numpy~=1.23.2
tortoise-orm>=0.19.2